Encoding of a ffnn added
