+++ dcgpy/CMakeLists.txt
@@ -52,4 +52,4 @@ LIBRARY DESTINATION ${DCGPY_INSTALL_PATH}
 )
 
 # Add the Python files.
-install(FILES __init__.py _version.py test.py _graphviz_visualize.py _sympy_simplify.py _visualize.py DESTINATION ${DCGPY_INSTALL_PATH})
+install(FILES __init__.py _version.py test.py _visualize.py _simplify.py _encode.py DESTINATION ${DCGPY_INSTALL_PATH})
