+++ dcgpy/__init__.py
@@ -2,7 +2,7 @@ __all__ = ['test']
 from .core import *
 
 # Add to expression a method to visualize the d-CGP graph
-from ._graphviz_visualize import _graphviz_visualize
+from ._visualize import _graphviz_visualize
 from ._visualize import _dcgpann_visualize
 expression_double.visualize = _graphviz_visualize
 expression_gdual_double.visualize = _graphviz_visualize
@@ -11,11 +11,14 @@ expression_ann_double.visualize = _dcgpann_visualize
 
 
 # Add to expression a method to simplify the d-CGP expression
-from ._sympy_simplify import _sympy_simplify, _ann_simplify
+from ._simplify import _sympy_simplify, _ann_simplify
 expression_double.simplify = _sympy_simplify
 expression_gdual_double.simplify = _sympy_simplify
 expression_gdual_vdouble.simplify = _sympy_simplify
 expression_ann_double.simplify = _ann_simplify
+
+# Add utilities
+from ._encode import encode_ffnn
 
 # Version setup.
 from ._version import __version__
