+++ dcgpy/_simplify.py
